+++ Tests/StencilTests/VariableSpec.swift
@@ -337,4 +337,18 @@ func testVariable() {
     }
 
   }
+
+  describe("inline if expression") {
+
+    $0.it("can conditionally render variable") {
+      let template: Template = "{{ variable if variable|uppercase == \"A\" }}"
+      try expect(template.render(Context(dictionary: ["variable": "a"]))) == "a"
+      try expect(template.render(Context(dictionary: ["variable": "b"]))) == ""
+    }
+
+    $0.it("throws when used invalid condition") {
+      let template: Template = "{{ variable if variable \"A\" }}"
+      try expect(template.render(Context(dictionary: ["variable": "a"]))).toThrow()
+    }
+  }
 }